loopback_latency: do not divide buffer_frames by 10.
The default buffer_frames is 480. 480/10=48. Setting
cb_threshold to 48 is too small and Cras will reject it.
The division by 10 was added four years ago. It was
an arbitrary value to get the buffer frames small.
Actually using 480 and not dividing it by 10 is more
reasonable.

BUG=chromium:779936
TEST=Run loopback_latency -c on eve and it does not show error.

Change-Id: I3d03fccc23e3695429748e2667b7f4683f234597
Reviewed-on: https://chromium-review.googlesource.com/753261
Commit-Ready: Wu-Cheng Li <wuchengli@chromium.org>
Tested-by: Wu-Cheng Li <wuchengli@chromium.org>
Reviewed-by: Chinyue Chen <chinyue@chromium.org>
Reviewed-by: Wu-Cheng Li <wuchengli@chromium.org>