alsa_conformance_test.py: Add rate and error criteria parameters
Make rate and error criteria can be set by user. It is useful for
if users want to run this test on different levels.

--rate-criteria-diff-pct
The pass criteria of rate. The value is a percentage of rate. For
example, 0.01 means the pass range is [47995.2, 48004.8] for rate
48000. (default: 0.01)

--rate-err-criteria
The pass criteria of rate error. (default: 10)

BUG=chromium:685077, b:135646043
TEST=alsa_conformace_test.py hw:0,0 PLAYBACK
--rate-criteria-diff-pct 0.1 --rate_err_criteria 5

Cq-Depend: chromium:1670769
Change-Id: Id121bfbd0b60590c2043a7b3f3e013316a8bd66c
Reviewed-on: https://chromium-review.googlesource.com/1670768
Tested-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Commit-Ready: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>