audiofuntest: Use nonblock write to avoid being stuck during playback
The buffer size of fd is small. If it is full, the playback will be
blocked until it has enough space. This behavior makes this thread not
able to be stopped immediately, which may cause the program stop to
capture audio and trigger overrun.

Make fd nonblocking so the thread can be interrupted at any time.

BUG=b:168281719
TEST=Pass audiofuntest on Dirinboz

Change-Id: Iee9f2d359a75c6286fa162ce33082fe5bbea9d69
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/audiotest/+/2457207
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Commit-Queue: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Tested-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>