alsa_conformance_test: Fix deadlock with Popen.wait()
replacing it with subprocess.run()

From the doc of Popen.wait():
Note This will deadlock when using stdout=PIPE or stderr=PIPE and
the child process generates enough output to a pipe such that it
blocks waiting for the OS pipe buffer to accept more data. Use
Popen.communicate() when using pipes to avoid that.

BUG=b:195267672
TEST=vm.Audio.ac97 (https://crrev.com/c/3119399)

Change-Id: I71f61cb004846d1ffb25f5bd33d411223d2d2f0f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/audiotest/+/3153077
Reviewed-by: Chih-Yang Hsia <paulhsia@chromium.org>
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Tested-by: Woody Chow <woodychow@google.com>
Commit-Queue: Woody Chow <woodychow@google.com>