audiofuntest: cover multiple frequency tests from low to high
Sometimes, we encounter some DUT that fail at low frequency test,
but because the picked frequency is random, it is possible that
this low frequency is not picked and the bad DUT test passes. So
cover multiple frequency tests from low to high, which may be an
improvement.

BUG=None
TEST=1. emerge-strongbad audiotest
     2. manually test on DUT

Change-Id: I3457184e216747263553db9b9477ae4f6a9a1862
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/audiotest/+/3379889
Tested-by: Ruihai Zhou <zhouruihai@huaqin.corp-partner.google.com>
Reviewed-by: Xuxin Xiong <xuxinxiong@huaqin.corp-partner.google.com>
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Reviewed-by: Simon Zhou <zhouguohui@huaqin.corp-partner.google.com>
Commit-Queue: Ruihai Zhou <zhouruihai@huaqin.corp-partner.google.com>