audiotest: update to _FORTIFY_SOURCE=3
Since this project explicitly specifies its FORTIFY level, update it to
the level CrOS wants to move to. This isn't _removed_, since it's copied
from a template file, and consistency with that seems useful.

The template file this was copied from was updated in
crrev.com/c/4735413.

BUG=b:288411201
TEST=CQ+1

Change-Id: Id71bcf8d4ea85f8a64637230680840b4d93b0ffd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/audiotest/+/4706998
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>