audiotest: loopback_latency: Allow test to use audio file
Allow loopback latency test to use audio file instead of sine wave so it
can measure the loopback latency when the noise cancellation is on.
For now it only supports raw pcm audio file that matches the format of
the playback for the test.

BUG=b:342317039
TEST=loopback_latency -i default -o default -f /tmp/record.raw

Change-Id: Icd45ee8109c39e2b1d2215e14ef4cb70056abfa3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/audiotest/+/5567812
Commit-Queue: Norman Bintang <normanbt@chromium.org>
Tested-by: Norman Bintang <normanbt@chromium.org>
Reviewed-by: Li-Yu Yu <aaronyu@google.com>