scripts: alsa_conformance_test: Add parameter for avail-delay option
The use of snd_pcm_avail_delay() in alsa_conformance_tool is made to be
optional and it is only enabled when --avail-delay parameter is passed.

While the use of snd_pcm_avail_delay() on systems where the driver does not
report delay is equal to the use of snd_pcm_avail(), make this optional
also via the python wrapper.

The use of snd_pcm_avail_delay() is required on systems where DSP offload
is used with large buffer to account for the delay caused by it.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@linux.intel.com>