scripts: alsa_conformance_test: Add parameter for block-size option
In case audio offloading is used with large buffer on the DSP side, the
DMA can move rapidly (burst) to move data between ALSA buffer and DSP.

With he default 240 buffer size this can cause xrun since the DMA might
move multiple times of the block-size.

Add parameter to the python tool which will pass the block-size to the tool
to make sure that the spinning on the buffer is not happening.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@linux.intel.com>